requirements.txt
django
gunicorn
openpyxl>=3.1.2
python-dotenv
psycopg2-binary
whitenoise
dj-database-url
djangorestframework
djangorestframework-simplejwt
django-cors-headers
groq>=0.11.0
requests>=2.31.0
Pillow